+++ src/main/java/com/sarasa/projectboard/controller/ArticleController.java
@@ -4,8 +4,8 @@ import com.sarasa.projectboard.domain.constant.FormStatus;
 import com.sarasa.projectboard.domain.type.SearchType;
 import com.sarasa.projectboard.dto.request.ArticleRequest;
 import com.sarasa.projectboard.dto.security.BoardPrincipal;
-import com.sarasa.projectboard.response.ArticleResponse;
-import com.sarasa.projectboard.response.ArticleWithCommentsResponse;
+import com.sarasa.projectboard.dto.response.ArticleResponse;
+import com.sarasa.projectboard.dto.response.ArticleWithCommentsResponse;
 import com.sarasa.projectboard.service.ArticleService;
 import com.sarasa.projectboard.service.PaginationService;
 import lombok.RequiredArgsConstructor;
